+++ src/drivers/cmakeServerDriver.ts
@@ -289,6 +289,10 @@ export class CMakeServerDriver extends CMakeDriver {
         return this.targets.reduce(targetReducer, []);
     }
 
+    get cmakeFiles(): string[] {
+        return this._cmakeInputFileSet.inputFiles.map(file => file.filePath);
+    }
+
     get generatorName(): string | null {
         return this._globalSettings ? this._globalSettings.generator : null;
     }
@@ -333,8 +337,8 @@ export class CMakeServerDriver extends CMakeDriver {
         await this._restartClient();
     }
 
-    doSetKit(need_clean: boolean, cb: () => Promise<void>): Promise<void> {
-        this._clientChangeInProgress = this._setKitAndRestart(need_clean, cb);
+    doSetKit(cb: () => Promise<void>): Promise<void> {
+        this._clientChangeInProgress = this._setKitAndRestart(false, cb);
         return this._clientChangeInProgress;
     }