+++ src/extension.ts
@@ -468,7 +468,7 @@ class ExtensionManager implements vscode.Disposable {
     }
 
     async configureExtensionInternal(trigger: ConfigureTrigger, cmt: CMakeTools): Promise<void> {
-        if (!await this.ensureActiveConfigurePresetOrKit(cmt)) {
+        if (trigger !== ConfigureTrigger.configureWithCache && !await this.ensureActiveConfigurePresetOrKit(cmt)) {
             return;
         }
 
@@ -598,7 +598,7 @@ class ExtensionManager implements vscode.Disposable {
                 if (result === configureButtonMessage) {
                     await this.configureExtensionInternal(ConfigureTrigger.buttonNewKitsDefinition, cmt);
                 } else {
-                    log.debug(localize('using.cache.to.configure.workspace.on.open', 'Using cache to configure workspace on open {0}', ws.uri.toString()));
+                    log.debug(localize('using.cache.to.configure.workspace.on.open', 'Attempting to use cache to configure workspace {0}', ws.uri.toString()));
                     await this.configureExtensionInternal(ConfigureTrigger.configureWithCache, cmt);
                 }
             }
@@ -703,7 +703,7 @@ class ExtensionManager implements vscode.Disposable {
                 log.debug(localize('update.intellisense.disabled', 'Not updating the configuration provider because {0} is set to {1}', '"C_Cpp.intelliSenseEngine"', '"Disabled"'));
                 return;
             }
-            if (!this.cppToolsAPI) {
+            if (!this.cppToolsAPI && !util.isTestMode()) {
                 try {
                     this.cppToolsAPI = await cpt.getCppToolsApi(cpt.Version.latest);
                 } catch (err) {
@@ -1098,10 +1098,7 @@ class ExtensionManager implements vscode.Disposable {
         return 0;
     }
 
-    mapCMakeToolsFolder(fn: CMakeToolsMapFn,
-        folder?: vscode.WorkspaceFolder,
-        precheck?: (cmt: CMakeTools) => Promise<boolean>,
-        cleanOutputChannel?: boolean): Promise<any> {
+    mapCMakeToolsFolder(fn: CMakeToolsMapFn, folder?: vscode.WorkspaceFolder, precheck?: (cmt: CMakeTools) => Promise<boolean>, cleanOutputChannel?: boolean): Promise<any> {
         if (cleanOutputChannel) {
             this.cleanOutputChannel();
         }