+++ test/extension-tests/single-root-UI/test/configure-and-build-presets.test.ts
@@ -8,7 +8,7 @@ import {
 import * as path from 'path';
 import * as vscode from 'vscode';
 
-suite('Build using Presets', async () => {
+suite('Build using Presets', () => {
     let testEnv: DefaultEnvironment;
     let compdb_cp_path: string;